Separate MovementSolver
